feat(repository): implementa metodo para agregar un Telefono de un Cliente

+++ src/main/java/com/misproyectos/interfaces/ClienteRepInterface.java
@@ -5,5 +5,5 @@ import com.misproyectos.models.Cliente;
 import java.sql.SQLException;
 
 public interface ClienteRepInterface {
-    int add(Cliente cliente) throws SQLException;
+    Long add(Cliente cliente) throws SQLException;
 }
